[ADD] Day and Night System

+++ include/field_player_avatar.h
@@ -4,7 +4,7 @@
 void PlayerStep(u8 direction, u16 newKeys, u16 heldKeys);
 void ClearPlayerAvatarInfo(void);
 void SetPlayerAvatarExtraStateTransition(u16 graphicsId, u8 transitionFlag);
-u8 GetPlayerAvatarGenderByGraphicsId(u16 gfxId);
+u16 GetPlayerAvatarGenderByGraphicsId(u16 gfxId);
 u8 TestPlayerAvatarFlags(u8 flag);
 u8 GetPlayerAvatarSpriteId(void);
 void PlayerGetDestCoords(s16 *x, s16 *y);
